Add the `aiida-sssp install` CLI command
This command will serve to automatically install a SSSP configuration by
downloading the corresponding archive from the Materials Cloud archive
and creating a new `SsspFamily` instance from it. Currently the
structure of the archives provided by the Materials Cloud are
inconsistent as some contain internal subdirectories and some don't. The
code currently does not fix this and will simply fail. Work is underway
to correct the actual archives on the Materials Cloud to respect a
strict consistent format.