Add md5 of download files to `SsspFamily` description
When installing an `SsspFamily` through `aiida-sssp` the md5 checksum of
the download pseudo archive and pseudo metadata file is added to the
description. This might become useful in the future if inconsistencies
are suspected for a given family and we want to make sure that the
downloaded files were at least correct.

Note that the description of the family is not the ideal location to
store this information, as it is mutable, but currently, `Group`
instances in `aiida-core` do not support arbitrary attributes to be
stored. For the time being this is the best solution.